Tyops

+++ packages/abi-typegen/src/abi/types/U16Type.ts
@@ -4,9 +4,9 @@ import type { IType } from '../../interfaces/IType';
 import { U8Type } from './U8Type';
 
 export class U16Type extends U8Type implements IType {
-  public static swayTypeExample = 'u8';
+  public static swayTypeExample = 'u16';
 
-  public name = 'u8';
+  public name = 'u16';
 
   public static MATCH_REGEX: RegExp = /^u16$/m;
